Update Keymap for windows & gaming support
Change some things around for adding better gaming support.

FMI: I noticed currently the second left key on the bottom row of the left half (which should be the first grey one) functions as CONTROL when held, but as KC_GRAVE (` & ~) when tapped. However this is not reflected in the keymap. Maybe I forgot to push changes while working at Todaytomorrow :(

(FMI stands for "For My Information")